requirements.txt
hubspot-api-client
requests
prettytable
ast